correct programming error in user
